Identifier: Extract it from `literals.rs`.
